Configure automatic deployment upon commits on main branch
This configures GitHub Actions to automatically deploy the changes approved by the committee to the server. Once approved, it basically instantiate a runner to ssh to the server and pull the changes to production.

Detailed documentation about how this file is configured can be found at: https://docs.github.com/en/free-pro-team@latest/actions/learn-github-actions/introduction-to-github-actions  

Before accepting this merge, please, make sure the GitHub Secrets have been configured with the necessary variables: `SERVER_IP`, `SERVER_USERNAME`, `SERVER_PASSWORD`, `PROJECT_PATH`. You can access that in the repository's Settings under Secrets. Secrets are environment variables that are encrypted and only exposed to selected actions. They are not passed to workflows that are triggered by a pull request from a fork, so it is safe to use.