update: view submission with new file names (#59)
* update: view submission with new file names

This updates the view page of a submission to also display the output files considering the new naming format. For instance, `ior_easy_read` was renamed to ìor-eay-read` in the latest version of the benchmark. This update will filter the files considering both names.

* update: include result.txt and io500.sh in filter

As requested, include the `result.txt` and `io500.sh` in the submission view.

* update: simplify list of files

Simplify list of files and let PHP transparently handle both naming conventions.