repo: change to team-based reviewer roles
Change CODEOWNERS to select patch reviewers based on a team
instead of individual users.  Unfortunately, while this can
automatically select reviewers for a particular page, it does
not affect how many reviewers are required for a page to be
merged.  That can only be configured on a per-branch or per-
repository basis, and needs to be set at the minimum number
of reviewers for any page landing to be approved.