Improve details of running and submission
Move the "Running" item before "Submission" in the page header,
since this is the logical order of progression for new users.

Update the running.php page to better describe the current process,
including specific examples of important parameters to be updated.
Mention that ior-rnd, mdworkbench, find-easy/hard can be ignored.

Remove redundant information on running.php and submission.php
to instead link to information on other available pages so it doesn't
need to be updated mulitple times (e.g. stale list of board members).